requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Machine Learning
torch>=2.0.0
gymnasium>=0.28.0

# Dashboard
streamlit>=1.28.0

# System monitoring
psutil>=5.9.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
scipy>=1.10.0

# Data handling
h5py>=3.8.0
